refactorizacion empleando patron de fabrica y modularizando el backend
