+++ backend/extensions.py
@@ -1,0 +1,8 @@
+# extensions.py
+from flask_sqlalchemy import SQLAlchemy
+from flask_bcrypt import Bcrypt
+from flask_jwt_extended import JWTManager
+
+db = SQLAlchemy()
+bcrypt = Bcrypt()
+jwt = JWTManager()
